==> 17_texture_boot_varImp.R <==
# Covariate importance plots for bootstrap texture models

library(ggplot2)
library(dplyr)
library(magrittr)
library(tidyr)

testn <- 14
mycrs <- "EPSG:25832"

# Fractions

fractions_alt  <- c("clay", "silt", "fine_sand", "coarse_sand", "SOC", "CaCO3")
fractions      <- fractions_alt
fraction_names <- c("Clay", "Silt", "Fine sand", "Coarse sand", "SOC", "CaCO3")

# Results folder

dir_code <- getwd()
root <- dirname(dir_code)
dir_dat <- paste0(root, "/digijord_data/")

dir_results <- dir_dat %>%
  paste0(., "/results_test_", testn, "/") %T>%
  dir.create()

dir_boot <- dir_results %>%
  paste0(., "/bootstrap/") %T>%
  dir.create()

# Covariate table

cov_cats <- dir_code %>%
  paste0(., "/cov_categories_20240305.csv") %>%
  read.table(
    sep = ",",
    header = TRUE
  )

l_cat <- cov_cats %>%
  mutate(
    Covariate = name,
    category = case_when(
      category == "basic" ~ scorpan,
      category == "WATEM" ~ "OR",
      category == "sentinel_composite" ~ "S2 time series",
      category == "bare_soil" ~ "Bare soil",
      desc_text %in% grep(
        "bare soil",
        cov_cats$desc_text,
        value = TRUE
      ) ~ "Bare soil", 
      .default = "Other"
    )
  ) %>%
  mutate(
    category = case_when(
      category == "RP" ~ "P",
      category == "CR" ~ "R",
      category == "OR" ~ "R",
      .default = category
    )
  )

l_cat %>% select(name, category)

# Covariate importance results

boot_imp_all <- dir_boot %>%
  paste0(., "/models_boot_importance_all.rds") %>%
  readRDS()

boot_imp_sand <- boot_imp_all %>%
  pivot_longer(
    cols = -c(fraction, Covariate),
    values_to = "Importance",
    names_to = "rep"
  ) %>%
  filter(
    fraction %in% c("clay", "silt", "coarse_sand")
  ) %>%
  group_by(Covariate, rep) %>%
  summarise(Importance = sum(Importance)/3) %>%
  ungroup() %>%
  arrange(-Importance) %>%
  mutate(fraction = "fine_sand")

l <- boot_imp_all %>%
  pivot_longer(
    cols = -c(fraction, Covariate),
    values_to = "Importance",
    names_to = "rep"
  ) %>%
  filter(fraction != "fine_sand") %>%
  bind_rows(boot_imp_sand)


# Join categories

l %<>%
  left_join(l_cat) %>%
  mutate(
    category = case_when(
      Covariate == "upper" ~ "Depth",
      Covariate == "lower" ~ "Depth",
      Covariate == "year" ~ "Time",
      Covariate == "SOM_removed" ~ "Method",
      category == "N" ~ "Spatial position",
      category == "R" ~ "Topography",
      category == "C" ~ "Climate",
      category == "C " ~ "Climate",
      category == "P" ~ "Parent materials",
      category == "S" ~ "Soil",
      category == "SO" ~ "Soil and organisms",
      category == "CR" ~ "Climate and topography",
      category == "OR" ~ "Organisms and topography",
      category == "O" ~ "Organisms",
      category == "RP" ~ "Parent materials",
      .default = category
    ),
    fraction = factor(
      fraction,
      levels = fractions,
      labels = fraction_names
    )
  )

cat_overall_order <- l %>%
  group_by(category) %>%
  summarise(
    mean = sum(Importance)
  ) %>%
  arrange(-mean) %>%
  select(category) %>%
  unlist() %>%
  as.character()

l %<>% mutate(
  category = factor(category, levels = rev(cat_overall_order))
)

levels(l$fraction) <- c(
  expression("Clay"),
  expression("Silt"),
  expression("Fine~sand"),
  expression("Coarse~sand"),
  expression("SOC"),
  expression("CaCO"[3])
)

# Colors for categories

library(colorRamps)
library(rcartocolor) # for colorblind palette

mycolors <- carto_pal(12, "Safe") %>% sort()

library(TSP)

catcolors <- l$category %>%
  levels() %>%
  length() %>%
  carto_pal(., "Safe")
names(catcolors) <- levels(l$category)
colScale <- scale_fill_manual(name = "category", values = catcolors)

# Make plot for categories

tiff(
  paste0(dir_results, "/boot_importance_categories_test_", testn, ".tiff"),
  width = 16,
  height = 10,
  units = "cm",
  res = 600
)

l %>%
  filter(Importance > 0) %>%
  group_by(category, fraction, rep) %>%
  summarise(Importance = sum(Importance)) %>%
  ungroup() %>%
  group_by(category, fraction) %>%
  mutate(mean = mean(Importance)) %>%
  ggplot(aes(x = category, y = Importance, fill = category)) +
  geom_violin(scale = "width") +
  facet_wrap(~ fraction, labeller = label_parsed) +
  coord_flip() +
  geom_point(aes(x = category, y = mean, fill = category)) +
  theme(
    text = element_text(family = "serif"),
    legend.position = "none"
    ) +
  xlab("Category")

try(dev.off())

# Plot importance for OGCs

ndir_plot <- 64

imp_ogc <- l %>%
  select(fraction, Covariate, rep, Importance) %>%
  filter(Covariate %in% grep('ogc_pi', Covariate, value = TRUE)) %>%
  arrange(Covariate) %>%
  mutate(
    dir = substr(
      Covariate,
      nchar(Covariate) - 2,
      nchar(Covariate)
    ) %>%
      as.numeric() %>%
      "*" (ndir_plot) %>%
      "/" (1000) %>%
      "+" (1) %>%
      round(digits = 0)
  ) %>%
  group_by(fraction, rep) %>%
  mutate(
    Importance = Importance/sum(Importance),
    Importance = case_when(
      !is.finite(Importance) ~ 0,
      .default = Importance
    )
  )

imp_ogc2 <- imp_ogc %>%
  mutate(dir = dir + ndir_plot)

imp_ogc %<>% bind_rows(., imp_ogc2)

brks <- seq(
  from = min(imp_ogc$dir),
  by = (max(imp_ogc$dir) + 1 - min(imp_ogc$dir))/4,
  length.out = 4
)

imp_ogc %>%
  ggplot(aes(x = dir, y = Importance)) +
  coord_polar(
    start = - pi/2 - pi/(ndir_plot*2),
    direction = -1) +
  geom_point() +
  facet_wrap(
    ~ fraction,
    ncol = 3
  ) +
  scale_x_continuous(
    breaks = brks,
    labels = c('E', 'N', 'W', 'S')
  ) +
  ylab('Covariate importance') +
  theme(
    axis.title.x = element_blank()
  )

tiff(
  paste0(dir_results, "/boot_importance_ogc_new_test", testn, ".tiff"),
  width = 16,
  height = 10,
  units = "cm",
  res = 600
)

imp_ogc %>%
  ungroup() %>%
  group_by(fraction, dir) %>%
  summarise(
    mean = mean(Importance),
    q = quantile(Importance, probs = 0.75)
  ) %>%
  ggplot(aes(x = dir, y = mean)) +
  coord_polar(
    start = - pi/2 - pi/(ndir_plot*2),
    direction = -1) +
  geom_linerange(aes(ymax = mean, ymin = 0), lineend = "round") +
  facet_wrap(
    ~ fraction,
    ncol = 3,
    labeller = label_parsed
  ) +
  scale_x_continuous(
    breaks = brks,
    labels = c('E', 'N', 'W', 'S')
  ) +
  ylab('Covariate importance') +
  theme(
    text = element_text(family = "serif"),
    axis.title.x = element_blank()
  )

try(dev.off())

# Top 3 most important covariates in each category, across all fractions

l %>%
  select(fraction, Covariate, rep, Importance, category) %>%
  filter(
    !(Covariate %in% grep('ogc_pi', Covariate, value = TRUE)),
    !(Covariate %in% c("upper", "lower", "SOM_removed"))
  ) %>%
  group_by(Covariate, category) %>%
  summarise(mean = mean(Importance)) %>%
  ungroup() %>%
  group_by(category) %>%
  arrange(-mean) %>%
  slice_head(n = 3)

# Top 10 environmental covariates overall

l %>%
  select(fraction, Covariate, rep, Importance, category) %>%
  filter(
    !(Covariate %in% grep('ogc_pi', Covariate, value = TRUE)),
    !(Covariate %in% c("upper", "lower", "SOM_removed")),
    fraction != "Fine~sand"
  ) %>%
  group_by(Covariate) %>%
  summarise(mean = mean(Importance)) %>%
  ungroup() %>%
  arrange(-mean) %>%
  slice_head(n = 10)

# Top 3 most important per fraction

l %>%
  select(fraction, Covariate, rep, Importance, category) %>%
  filter(
    !(Covariate %in% grep('ogc_pi', Covariate, value = TRUE)),
    !(Covariate %in% c("upper", "lower", "SOM_removed"))
  ) %>%
  group_by(Covariate, fraction) %>%
  summarise(mean = mean(Importance)) %>%
  ungroup() %>%
  group_by(fraction) %>%
  arrange(-mean) %>%
  slice_head(n = 3)

# Top 2 per fraction + category

l %>%
  select(fraction, Covariate, rep, Importance, category) %>%
  filter(
    !(Covariate %in% grep('ogc_pi', Covariate, value = TRUE)),
    !(Covariate %in% c("upper", "lower", "SOM_removed"))
  ) %>%
  group_by(Covariate, fraction, category) %>%
  summarise(mean = mean(Importance)) %>%
  ungroup() %>%
  group_by(fraction, category) %>%
  arrange(-mean) %>%
  slice_head(n = 2) %>%
  mutate(
    fraction = make.names(fraction),
    category = make.names(category)
    ) %>%
  print(n = 100)

# Top 10 most important per fraction

l %>%
  select(fraction, Covariate, rep, Importance, category) %>%
  filter(
    !(Covariate %in% grep('ogc_pi', Covariate, value = TRUE)),
    !(Covariate %in% c("upper", "lower", "SOM_removed"))
  ) %>%
  group_by(Covariate, fraction) %>%
  summarise(mean = mean(Importance)) %>%
  ungroup() %>%
  group_by(fraction) %>%
  arrange(-mean) %>%
  slice_head(n = 10) %>%
  mutate(
    fraction = make.names(fraction)
  ) %>%
  print(n = 100)


# Covariate importance for time series

s2_bandnames <- c("b2", "b3", "b4", "b5", "b6", "b7", "b8", "b8a", "b11", "b12")
s2_bandlabels <- c("2", "3", "4", "5", "6", "7", "8", "8a", "11", "12")

s2_time_imp <- l %>%
  filter(
    category == "S2 time series",
    fraction != "Fine~sand"
    ) %>%
  group_by(fraction, Covariate) %>%
  summarise(Importance = sum(Importance)/100) %>%
  mutate(
    Band = substr(
      Covariate,
      32,
      nchar(Covariate)
    ),
    Band = factor(
      Band,
      levels = s2_bandnames,
      labels = s2_bandlabels
      ),
    Time = substr(
      Covariate,
      14,
      21
    ),
    Time = factor(Time, labels = c("Spring", "June", "July"))
  ) %>%
  group_by(fraction) %>%
  mutate(Importance = Importance/sum(Importance)) %>%
  ungroup()

tiff(
  paste0(dir_results, "/boot_importance_s2_test", testn, ".tiff"),
  width = 16,
  height = 10,
  units = "cm",
  res = 600
)

s2_time_imp %>%
  ggplot(
    aes(
      x = Importance,
      y = Band,
      colour = Time,
      shape = Time,
      group = Time,
      linetype = Time
      )
    ) +
  geom_point() +
  geom_line(orientation = "y") +
  facet_wrap(
    ~ fraction,
    nrow = 1,
    labeller = label_parsed) +
  scale_y_discrete(limits = rev) +
  theme(
    text = element_text(family = "serif")
  )

try(dev.off())

# General covariate importance table

l %>%
  mutate(fraction = factor(fraction, labels = fraction_names)) %>%
  filter(fraction != "Fine sand") %>%
  select(fraction, Covariate, Importance) %>%
  group_by(fraction, Covariate) %>%
  summarise(
    Importance = sum(Importance)/100
  ) %>%
  filter(Importance > 0) %>%
  group_by(fraction) %>%
  arrange(-Importance) %>%
  slice_head(n = 20) %>%
  write.table(
    paste0(dir_results, "/varImp_top20.csv"),
    sep = ";",
    row.names = FALSE
  )

# Mean importance for all covariates

l %>%
  mutate(fraction = factor(fraction, labels = fraction_names)) %>%
  filter(fraction != "Fine sand") %>%
  select(fraction, Covariate, Importance) %>%
  group_by(fraction, Covariate) %>%
  summarise(
    Importance = sum(Importance)/100
  ) %>%
  filter(Importance > 0) %>%
  group_by(fraction) %>%
  arrange(-Importance) %>%
  pivot_wider(
    values_from = Importance,
    names_from = fraction
    ) %>%
  mutate(mean = rowMeans(pick(where(is.numeric)))) %>%
  arrange(-mean) %>%
  as.data.frame() %>%
  write.table(
    paste0(dir_results, "/varImp_all.csv"),
    sep = ";",
    row.names = FALSE
  )

# END
